[SD-193] Abstract simple db actions
